+++ supply-chain-testkit/src/test/java/io/github/joshuamatosdev/security/supplychain/testkit/ActionPinPolicyContractTest.java
@@ -1,0 +1,11 @@
+package io.github.joshuamatosdev.security.supplychain.testkit;
+
+import io.github.joshuamatosdev.security.supplychain.policy.ActionPinPolicy;
+
+class ActionPinPolicyContractTest implements ActionPinPolicyContract {
+
+    @Override
+    public ActionPinPolicy policy() {
+        return new ActionPinPolicy();
+    }
+}